+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 
 - Do not encode payloads in cassettes as base64 by default. This change makes Schemathesis match the default Ruby's VCR behavior and
   leads to more human-readable cassettes. Use ``--cassette-preserve-exact-body-bytes`` to restore the old behavior. `#1413`_
+- Bump ``hypothesis-graphql`` to ``0.7.0``.
 
 .. _v3.14.2: